backend-student-portal/requirements.txt
python-dotenv
flask
flask-cors
openai
psycopg2-binary
flask-sqlalchemy
flask-bcrypt
flask-jwt-extended
flask-mail